# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  internal_handlers.createOIDCRequest:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      display_name:
        type: string
      issuer_url:
        type: string
      name:
        type: string
      scopes:
        type: string
    type: object
...
  internal_handlers.oidcProviderResponse:
    properties:
      auto_provision:
        type: boolean
      client_id:
        type: string
      created_at:
        type: string
      default_group_id:
        type: integer
      display_name:
        type: string
      extra_authorize_params:
        additionalProperties:
          type: string
        type: object
      id:
        type: integer
      is_enabled:
        type: boolean
      issuer_url:
        type: string
      name:
        type: string
      scopes:
        type: string
      updated_at:
        type: string
    type: object
...
  internal_handlers.updateOIDCRequest:
    properties:
      client_id:
        type: string
      client_secret:
        description: empty = keep existing
        type: string
      display_name:
        type: string
      is_enabled:
        type: boolean
      issuer_url:
        type: string
      name:
        type: string
      scopes:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise gateway for Autonomi decentralized storage. Provides file
    upload/download, user management, API tokens, webhooks, and admin controls.
  title: Indelible API
  version: "2.0"
paths:
...
  /admin/oidc/providers:
    get:
      description: Return all configured OIDC identity providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_handlers.oidcProviderResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List OIDC providers
      tags:
      - 'Admin: OIDC'
    post:
      consumes:
      - application/json
      description: Register a new OIDC identity provider with encrypted client secret
      parameters:
      - description: OIDC provider details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createOIDCRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.oidcProviderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an OIDC provider
      tags:
      - 'Admin: OIDC'
  /admin/oidc/providers/{id}:
    delete:
      description: Remove an OIDC identity provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an OIDC provider
      tags:
      - 'Admin: OIDC'
    put:
      consumes:
      - application/json
      description: Update an existing OIDC provider's configuration (empty client_secret
        keeps existing)
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated provider fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateOIDCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.oidcProviderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an OIDC provider
      tags:
      - 'Admin: OIDC'
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT or API token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"